Rename operator folder to physical_plan_v2
