+++ src/execution/volcano_executor/filter.rs
@@ -4,7 +4,7 @@ use crate::binder::expression::BoundExpression;
 use crate::dbtype::value::Value;
 use crate::execution::execution_plan::ExecutionPlan;
 use crate::{
-    execution::ExecutionContext, optimizer::operator::PhysicalPlanV2, storage::tuple::Tuple,
+    execution::ExecutionContext, optimizer::physical_plan_v2::PhysicalPlanV2, storage::tuple::Tuple,
 };
 use std::sync::Arc;
