+++ src/execution/volcano_executor/insert.rs
@@ -4,7 +4,7 @@ use crate::{
     catalog::column::Column,
     dbtype::value::Value,
     execution::{execution_plan::ExecutionPlan, ExecutionContext},
-    optimizer::operator::PhysicalPlanV2,
+    optimizer::physical_plan_v2::PhysicalPlanV2,
     storage::{
         table_heap,
         tuple::{Tuple, TupleMeta},
