+++ src/execution/volcano_executor/limit.rs
@@ -2,7 +2,7 @@ use std::sync::{atomic::AtomicUsize, Arc, Mutex};
 
 use crate::{
     execution::{execution_plan::ExecutionPlan, ExecutionContext},
-    optimizer::{operator::PhysicalPlanV2, physical_plan::PhysicalPlan},
+    optimizer::{physical_plan::PhysicalPlan, physical_plan_v2::PhysicalPlanV2},
     storage::tuple::Tuple,
 };
