+++ src/execution/volcano_executor/project.rs
@@ -1,6 +1,6 @@
 use crate::execution::execution_plan::ExecutionPlan;
 use crate::{
-    execution::ExecutionContext, optimizer::operator::PhysicalPlanV2, storage::tuple::Tuple,
+    execution::ExecutionContext, optimizer::physical_plan_v2::PhysicalPlanV2, storage::tuple::Tuple,
 };
 use std::sync::Arc;
