+++ src/optimizer/physical_plan_v2/insert.rs
